+++ lib/gateway/gatewayServer.js
@@ -22,9 +22,7 @@ function getRealIp(req /*: express$Request */) {
   if (realIp && realIp.indexOf(',') > -1) {
     realIp = realIp.split(',')[0]
   }
-
   const ipv4 = realIp.substring(realIp.lastIndexOf(':') + 1, realIp.length)
-
   return ipv4
 }
