+++ lib/shared/config.js
@@ -96,4 +96,8 @@ const config = {
   parseUris
 }
 
+const [KubeSailApiTarget, KubeSailApiPort] = config.KUBESAIL_API_TARGET.split(':')[0]
+config.apiTarget = KubeSailApiTarget
+config.apiPort = KubeSailApiPort
+
 module.exports = config
