+++ package.json
@@ -40,6 +40,7 @@
     "http-headers": "^3.0.2",
     "ioredis": "^4.16.3",
     "kubernetes-client": "^9.0.0",
+    "lru-cache": "^5.1.1",
     "mocha": "^7.1.2",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.3",
